requirements.txt
anyio==4.7.0
attrs==24.3.0
base58==2.1.1
borsh-construct==0.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
construct==2.10.68
construct-typing==0.5.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jsonalias==0.1.1
numpy==2.2.1
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
solana==0.36.2
solders==0.23.0
sumtypes==0.1a6
tabulate==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
websockets==13.1
